requirements.txt
boto3==1.35.72
pydantic==2.10.2
python-dotenv==1.0.1